requirements/base.txt
-c constraints.txt

tapad-common==12.0.0
apache-airflow[google]==2.4.3